requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
boto3==1.38.19
botocore==1.38.19
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2
django-cors-headers==4.7.0
django-storages==1.14.6
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flower==2.0.1
humanize==4.12.3
jmespath==1.0.1
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.2.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
